refactor: update promotion grid section
