+++ app/sections/PromotionGrid/PromotionGrid.tsx
