+++ app/weaverse/components.ts
@@ -24,8 +24,8 @@ import * as Main from '~/sections/main';
 import * as Map from '~/sections/map';
 import * as Page from '~/sections/page';
 import * as ProductInformation from '~/sections/product-information';
-import * as PromotionGridItem from '~/sections/promotion-grid/item';
-import * as PromotionGrid from '~/sections/promotion-grid/promotion-grid';
+import * as PromotionGridItem from '~/sections/PromotionGrid/item';
+import * as PromotionGrid from '~/sections/PromotionGrid/PromotionGrid';
 import * as RelatedArticles from '~/sections/related-articles';
 import * as RelatedProducts from '~/sections/related-products';
 import * as RichTextButtonItem from '~/sections/RichText/Button';
